fix: move model loading functions to model_registry; update imports in services
