+++ ispitch-api/src/core/model_registry.py
@@ -1,0 +1,14 @@
+from functools import lru_cache
+
+import spacy
+import whisper
+
+
+@lru_cache(maxsize=1)
+def get_whisper_model():
+    return whisper.load_model('base')
+
+
+@lru_cache(maxsize=1)
+def get_spacy_model():
+    return spacy.load('pt_core_news_sm')
